+++ iOS/SwimCoach/Navigation/AppRouter.swift
@@ -26,4 +26,5 @@ enum AppDestination: Hashable {
     case analyzing(URL)
     case results(AnalysisResult)
     case history
+    case compare(earlier: AnalysisResult, later: AnalysisResult)
 }